StrinGhost

+++ install.sh
@@ -12,11 +12,21 @@ RED=$(tput setaf 203 bold)
 CYAN=$(tput setaf 6 bold)
 MAUVE="\e[38;2;203;166;247;1m"
 
+LOG_FILE='kalighost.out'
+BIN_PATH='/usr/local/bin'
+
+
+if [[ $1 == "--no-picom" ]]; then
+  PICOM=false
+else
+  PICOM=true
+fi
+
 # Check if running as root. If root, script will exit
 if [[ $EUID -eq 0 ]]; then
     echo "${ERROR}" "This script should not be executed as root! Exiting.......${RESET}"
     exit 1
-fi
+read -r 
 
 clear
 
@@ -37,7 +47,7 @@ echo "${RED}Welcome to StrinGhost's KaliGhost Install Script!${RESET}"
 echo
 echo -e "${RED}ATTENTION: ${MAUVE}Run a full system update and Reboot first!! (Highly Recommended) ${RESET}"
 echo
-echo -e "${RED}NOTE: ${CYAN}After installation, all outputs and errors are saved to 'kalighost.out'. Please review it if any issues arise!"
+echo -e "${RED}NOTE: ${CYAN}After installation, all outputs and errors are saved to '$LOG_FILE'. Please review it if any issues arise!"
 echo
 echo -e "${RED}NOTE: ${CYAN}If you are installing on a VM, ensure to enable 3D acceleration else you may face issues!"
 echo
@@ -54,7 +64,7 @@ colorize_prompt() {
     local color="$1"
     local message="$2"
     echo -n "${color} ${message}${RESET}"
-    echo -e "\n${color}${message}${RESET}\n" &>> kalighost.out
+    echo -e "\n${color}${message}${RESET}\n" &>> "$LOG_FILE"
     printf "\n%.0s" {1..2}
 }
 
@@ -67,12 +77,12 @@ cmd_exec() {
     if [[ $? -ne 0 ]]; then
         echo "$(colorize_prompt "${ERROR}"  "Error: Command '$cmd' failed to execute.")" 
     fi
-    echo -e "$output\n\n" &>> kalighost.out
+    echo -e "$output\n\n" &>> "$LOG_FILE"
 }
 
 #######################################################################
 
-echo -e "$WELCOME\n\n\n" > kalighost.out
+echo -e "$WELCOME\n\n\n" > "$LOG_FILE"
 # Check for sudo permission
 if sudo -n true 2>/dev/null; then
     echo "$(colorize_prompt "${OK}"  "Sudo privileges detected.")"
@@ -86,7 +96,7 @@ fi
 echo "$(colorize_prompt "${CAT}"  "Installing Packages and Dependencies. This may take a while depending on your internet speed...")" 
 
 cmd_exec sudo apt update -y
-cmd_exec sudo apt-get install -y zsh python3 npm arandr flameshot arc-theme feh i3 polybar lxappearance python3-pip curl rofi kitty unclutter-xfixes unzip betterlockscreen wtmpdb bat eza net-tools fzf tmux btop ripgrep zoxide xclip gh sassc ncdu
+cmd_exec sudo apt-get install -y zsh python3 npm arandr flameshot arc-theme feh i3 polybar lxappearance python3-pip curl rofi kitty unclutter-xfixes unzip betterlockscreen wtmpdb bat eza net-tools fzf tmux btop ripgrep zoxide xclip gh sassc wget curl
 cmd_exec sudo apt-get install -y libxcb-shape0-dev libxcb-keysyms1-dev libpango1.0-dev libxcb-util0-dev libxcb1-dev libxcb-icccm4-dev libyajl-dev libev-dev libxcb-xkb-dev libxcb-cursor-dev libxkbcommon-dev libxcb-xinerama0-dev libxkbcommon-x11-dev libstartup-notification0-dev libxcb-randr0-dev libxcb-xrm0 libxcb-xrm-dev autoconf meson libxcb-render-util0-dev libxcb-shape0-dev libxcb-xfixes0-dev
 echo "$(colorize_prompt "${OK}"  "Packages and Dependencies Installed.")" 
 
@@ -118,19 +128,21 @@ fi
 # Copy configuration files
 echo "$(colorize_prompt "${CAT}"  "Setting up DotFiles.")" 
 cp -r config/* ~/.config/
-chmod +x chbg
 cp chbg ~/.chbg
 cp -r images ~/.images
-sudo cp $(which batcat) $(which cat)
-sudo cp $(which eza) $(which ls)
-sudo cp bin/picom /usr/bin/picom
-sudo cp bin/copy /usr/bin/copy
+sudo cp bin/copy "$BIN_PATH/"
+if [[ "$PICOM" = "false" ]]; then
+  cmd_exec sudo apt install picom # old version of picom
+else
+  sudo cp bin/picom "$BIN_PATH/"
+fi
 
 # Make scripts executable
-sudo chmod +x ~/.config/polybar/launch.sh 
-sudo chmod +x ~/.config/rofi/app-launcher/launch.sh
-sudo chmod +x ~/.config/rofi/clipboard/launch.sh
-sudo chmod +x ~/.config/rofi/power/launch.sh
+chmod +x ~/.chbg
+chmod +x ~/.config/polybar/launch.sh 
+chmod +x ~/.config/rofi/app-launcher/launch.sh
+chmod +x ~/.config/rofi/clipboard/launch.sh
+chmod +x ~/.config/rofi/power/launch.sh
 
 
 # Configure Betterlockscreen 
@@ -147,19 +159,19 @@ if [ ! -d "$HOME/.tmux/plugins/tpm" ]; then
 fi
 
 # GreenClip
-if [ ! -f "/usr/bin/greenclip" ]; then
+if [ ! -f "$BIN_PATH/greenclip" ]; then
     echo "$(colorize_prompt "${CAT}"  "Installing GreenClip")" 
     cmd_exec wget -q https://github.com/erebe/greenclip/releases/download/v4.2/greenclip 
     chmod +x greenclip
-    sudo mv greenclip /usr/bin/greenclip
+    sudo mv greenclip "$BIN_PATH/greenclip"
 fi
 
 # Install Neovim with Nvchad
-if [ ! -f "/usr/bin/nvim" ]; then
+if [ ! -f "$BIN_PATH/nvim" ]; then
    echo "$(colorize_prompt "${CAT}"  "Installing NeoVim.")" 
    cmd_exec wget -q https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz -O nvim-linux64.tar.gz
    tar -xzf nvim-linux64.tar.gz
-   sudo cp -r nvim-linux64/* /usr/
+   sudo cp -r nvim-linux64/* /usr/local/
    rm -r nvim-linux64.tar.gz nvim-linux64
    echo "$(colorize_prompt "${CAT}"  "Installing NvChad Starter.")" 
    rm -rf ~/.config/nvim
@@ -167,9 +179,9 @@ if [ ! -f "/usr/bin/nvim" ]; then
 fi
 
 # Install Oh My Posh
-if [ ! -f "/usr/bin/oh-my-posh" ]; then
+if [ ! -f "$BIN_PATH/oh-my-posh" ]; then
   echo "$(colorize_prompt "${CAT}"  "Setting Up Zsh.")" 
-  curl -s https://ohmyposh.dev/install.sh | sudo bash -s -- -d /usr/bin/ 
+  curl -s https://ohmyposh.dev/install.sh | sudo bash -s -- -d "$BIN_PATH/" 
   cp zshrc ~/.zshrc
 fi
 
@@ -181,18 +193,19 @@ clear
 echo "$WELCOME"
 printf "\n%.0s" {1..2}  
 echo -e "$(colorize_prompt "${MAUVE}"  "(+) KaliGhost Installation Completed!")" 
-echo -e "$(colorize_prompt "${WARN}"  "all outputs and errors are saved to 'kalighost.out'. Please review it if any issues arise!")" 
+echo -e "$(colorize_prompt "${WARN}"  "all outputs and errors are saved to '$LOG_FILE'. Please review it if any issues arise!")" 
 
 printf "\n%.0s" {1..1}
 read -rp "${CAT} Would you like to install Additional Wallpapers? (y/n): " WALL 
 if [[ "$WALL" =~ ^[Yy]$ ]]; then
-  git config http.postBuffer 524288000 # increase Buffer Size
-  cmd_exec git clone https://github.com/StrinGhost/Wallpaper-Collection.git --depth 1 
-  cp -r Wallpaper-Collection/* "$HOME/.images/"
-  rm -r Wallpaper-Collection
-  echo $(colorize_prompt "${OK}" "Installed Wallpapers Successfully.")
+  if git clone "https://github.com/StrinGhost/Wallpaper-Collection.git" --depth 1; then 
+      cp -r Wallpaper-Collection/* "$HOME/.images/"
+      rm -r Wallpaper-Collection
+      echo "$(colorize_prompt "${OK}" "Installed Wallpapers Successfully.")"
+  else
+      echo "$(colorize_prompt "${ERROR}" "Failed to Clone the Github Repo: 'https://github.com/StrinGhost/Wallpaper-Collection'. Please Manually Install!")"
+  fi
 fi
-
 
 # Prompt to Reboot
 printf "\n%.0s" {1..1}
@@ -202,3 +215,5 @@ if [[ "$ANS" =~ ^[Yy]$ ]]; then
 else
     echo "$(colorize_prompt "${MAUVE}"  "Configuration Complete: Manual Reboot Required.")" 
 fi
+
+
